Fix false outdated download for album zip

+++ npgCore/extensions/downloadList/download_statistics.php
@@ -58,7 +58,7 @@ function printBarGraph() {
 				$countlines++;
 			}
 			$outdated = '';
-			if (!file_exists(internalToFilesystem($item['aux']))) {
+			if (!file_exists(internalToFilesystem($item['aux'])) && !file_exists(ALBUM_FOLDER_SERVERPATH . stripSuffix($item['aux']))) {
 				$outdated = ' class="unpublished_item"';
 			}
 			?>
@@ -68,7 +68,13 @@ function printBarGraph() {
 				</td>
 				<td class="statistic_title" <?php echo $style; ?>>
 					<strong<?php echo $outdated; ?>>
-						<?php echo html_encode($item['aux']); ?>
+						<?php
+						$album = "";
+						if (strpos($item['aux'], SERVERPATH) === false) { // it's an album
+							$album = " [" . gettext('Album') . "]";
+						}
+						echo html_encode($item['aux'] . $album);
+						?>
 					</strong>
 				</td>
 				<td class="statistic_graphwrap" <?php echo $style; ?>>
@@ -104,7 +110,7 @@ echo '</head>';
 					$result = query_full_array($sql);
 					if ($result) {
 						foreach ($result as $row) {
-							if (!file_exists(internalToFilesystem($row['aux']))) {
+							if (!file_exists(internalToFilesystem($row['aux'])) && !file_exists(ALBUM_FOLDER_SERVERPATH . stripSuffix($row['aux']))) {
 								query('DELETE FROM ' . prefix('plugin_storage') . ' WHERE `id`=' . $row['id']);
 							}
 						}
